+++ tests/classification/test_average_precision.py
@@ -30,7 +30,7 @@ from torchmetrics.functional import average_precision
 seed_all(42)
 
 
-def _sk_average_precision_score(y_true, probas_pred, num_classes=1):
+def _sk_average_precision_score(y_true, probas_pred, num_classes=1, average=None):
     if num_classes == 1:
         return sk_average_precision_score(y_true, probas_pred)
 
@@ -39,33 +39,41 @@ def _sk_average_precision_score(y_true, probas_pred, num_classes=1):
         y_true_temp = np.zeros_like(y_true)
         y_true_temp[y_true == i] = 1
         res.append(sk_average_precision_score(y_true_temp, probas_pred[:, i]))
+
+    if average == 'macro':
+        return np.array(res).mean()
+    elif average == 'weighted':
+        weights = np.bincount(y_true) if y_true.max() > 1 else y_true.sum(axis=0)
+        weights = weights / sum(weights)
+        return (np.array(res) * weights).sum()
+
     return res
 
 
-def _sk_avg_prec_binary_prob(preds, target, num_classes=1):
+def _sk_avg_prec_binary_prob(preds, target, num_classes=1, average=None):
     sk_preds = preds.view(-1).numpy()
     sk_target = target.view(-1).numpy()
 
-    return _sk_average_precision_score(y_true=sk_target, probas_pred=sk_preds, num_classes=num_classes)
+    return _sk_average_precision_score(y_true=sk_target, probas_pred=sk_preds, num_classes=num_classes, average=average)
 
 
-def _sk_avg_prec_multiclass_prob(preds, target, num_classes=1):
+def _sk_avg_prec_multiclass_prob(preds, target, num_classes=1, average=None):
     sk_preds = preds.reshape(-1, num_classes).numpy()
     sk_target = target.view(-1).numpy()
 
-    return _sk_average_precision_score(y_true=sk_target, probas_pred=sk_preds, num_classes=num_classes)
+    return _sk_average_precision_score(y_true=sk_target, probas_pred=sk_preds, num_classes=num_classes, average=average)
 
 
-def _sk_avg_prec_multilabel_prob(preds, target, num_classes):
+def _sk_avg_prec_multilabel_prob(preds, target, num_classes=1, average=None):
     sk_preds = preds.reshape(-1, num_classes).numpy()
     sk_target = target.view(-1, num_classes).numpy()
-    return sk_average_precision_score(sk_target, sk_preds, average=None)
+    return sk_average_precision_score(sk_target, sk_preds, average=average)
 
 
-def _sk_avg_prec_multidim_multiclass_prob(preds, target, num_classes=1):
+def _sk_avg_prec_multidim_multiclass_prob(preds, target, num_classes=1, average=None):
     sk_preds = preds.transpose(0, 1).reshape(num_classes, -1).transpose(0, 1).numpy()
     sk_target = target.view(-1).numpy()
-    return _sk_average_precision_score(y_true=sk_target, probas_pred=sk_preds, num_classes=num_classes)
+    return _sk_average_precision_score(y_true=sk_target, probas_pred=sk_preds, num_classes=num_classes, average=average)
 
 
 @pytest.mark.parametrize(
@@ -77,30 +85,37 @@ def _sk_avg_prec_multidim_multiclass_prob(preds, target, num_classes=1):
         (_input_multilabel.preds, _input_multilabel.target, _sk_avg_prec_multilabel_prob, NUM_CLASSES),
     ],
 )
+@pytest.mark.parametrize("average", ['micro', 'macro', 'weighted', None])
 class TestAveragePrecision(MetricTester):
     @pytest.mark.parametrize("ddp", [True, False])
     @pytest.mark.parametrize("dist_sync_on_step", [True, False])
-    def test_average_precision(self, preds, target, sk_metric, num_classes, ddp, dist_sync_on_step):
+    def test_average_precision(self, preds, target, sk_metric, num_classes, average, ddp, dist_sync_on_step):
+        if target.max() > 1 and average == 'micro':
+            pytest.skip('average=micro and multiclass input cannot be used together')
+
         self.run_class_metric_test(
             ddp=ddp,
             preds=preds,
             target=target,
             metric_class=AveragePrecision,
-            sk_metric=partial(sk_metric, num_classes=num_classes),
+            sk_metric=partial(sk_metric, num_classes=num_classes, average=average),
             dist_sync_on_step=dist_sync_on_step,
-            metric_args={"num_classes": num_classes},
+            metric_args={"num_classes": num_classes, 'average': average},
         )
 
-    def test_average_precision_functional(self, preds, target, sk_metric, num_classes):
+    def test_average_precision_functional(self, preds, target, sk_metric, num_classes, average):
+        if target.max() > 1 and average == 'micro':
+            pytest.skip('average=micro and multiclass input cannot be used together')
+
         self.run_functional_metric_test(
             preds=preds,
             target=target,
             metric_functional=average_precision,
-            sk_metric=partial(sk_metric, num_classes=num_classes),
-            metric_args={"num_classes": num_classes},
+            sk_metric=partial(sk_metric, num_classes=num_classes, average=average),
+            metric_args={"num_classes": num_classes, 'average': average},
         )
 
-    def test_average_precision_differentiability(self, preds, sk_metric, target, num_classes):
+    def test_average_precision_differentiability(self, preds, sk_metric, target, num_classes, average):
         self.run_differentiability_test(
             preds=preds,
             target=target,
@@ -126,3 +141,26 @@ class TestAveragePrecision(MetricTester):
 )
 def test_average_precision(scores, target, expected_score):
     assert average_precision(scores, target) == expected_score
+
+
+def test_average_precision_warnings_and_errors():
+    """ Test that the correct errors and warnings gets raised """
+
+    # check average argument
+    with pytest.raises(ValueError, match='Expected argument `average` to be one .*'):
+        AveragePrecision(num_classes=5, average='samples')
+
+    # check that micro average cannot be used with multilabel input
+    pred = tensor([[0.75, 0.05, 0.05, 0.05, 0.05],
+                   [0.05, 0.75, 0.05, 0.05, 0.05],
+                   [0.05, 0.05, 0.75, 0.05, 0.05],
+                   [0.05, 0.05, 0.05, 0.75, 0.05]])
+    target = tensor([0, 1, 3, 2])
+    average_precision = AveragePrecision(num_classes=5, average='micro')
+    with pytest.raises(ValueError, match='Cannot use `micro` average with multi-class input'):
+        average_precision(pred, target)
+
+    # check that warning is thrown when average=macro and nan is encoutered in individual scores
+    average_precision = AveragePrecision(num_classes=5, average='macro')
+    with pytest.warns(UserWarning, match='Average precision score for one or more classes was `nan`.*'):
+        average_precision(pred, target)