+++ torchmetrics/classification/average_precision.py
@@ -44,6 +44,21 @@ class AveragePrecision(Metric):
             which for binary problem is translate to 1. For multiclass problems
             this argument should not be set as we iteratively change it in the
             range [0,num_classes-1]
+        average:
+            defines the reduction that is applied in the case of multiclass and multilabel input. 
+            Should be one of the following:
+
+            - ``'macro'`` [default]: Calculate the metric for each class separately, and average the
+              metrics across classes (with equal weights for each class).
+            - ``'micro'``: Calculate the metric globally, across all samples and classes. Cannot be
+              used with multiclass input.
+            - ``'weighted'``: Calculate the metric for each class separately, and average the
+              metrics across classes, weighting each class by its support.
+            - ``'none'`` or ``None``: Calculate the metric for each class separately, and return
+              the metric for every class.
+            - ``'samples'``: Calculate the metric for each sample, and average the metrics
+              across samples (with equal weights for each sample).
+
         compute_on_step:
             Forward only calls ``update()`` and return None if this is set to False. default: True
         dist_sync_on_step:
@@ -66,7 +81,7 @@ class AveragePrecision(Metric):
         ...                      [0.05, 0.05, 0.75, 0.05, 0.05],
         ...                      [0.05, 0.05, 0.05, 0.75, 0.05]])
         >>> target = torch.tensor([0, 1, 3, 2])
-        >>> average_precision = AveragePrecision(num_classes=5)
+        >>> average_precision = AveragePrecision(num_classes=5, average=None)
         >>> average_precision(pred, target)
         [tensor(1.), tensor(1.), tensor(0.2500), tensor(0.2500), tensor(nan)]
     """
@@ -78,6 +93,7 @@ class AveragePrecision(Metric):
         self,
         num_classes: Optional[int] = None,
         pos_label: Optional[int] = None,
+        average: Optional[str] = 'macro',
         compute_on_step: bool = True,
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
@@ -90,6 +106,11 @@ class AveragePrecision(Metric):
 
         self.num_classes = num_classes
         self.pos_label = pos_label
+        allowed_average = ('micro', 'macro', 'weighted', None)
+        if average not in allowed_average:
+            raise ValueError(f'Expected argument `average` to be one of {allowed_average}'
+                             f' but got {average}')
+        self.average = average
 
         self.add_state("preds", default=[], dist_reduce_fx="cat")
         self.add_state("target", default=[], dist_reduce_fx="cat")
@@ -107,7 +128,7 @@ class AveragePrecision(Metric):
             target: Ground truth values
         """
         preds, target, num_classes, pos_label = _average_precision_update(
-            preds, target, self.num_classes, self.pos_label
+            preds, target, self.num_classes, self.pos_label, self.average
         )
         self.preds.append(preds)
         self.target.append(target)
@@ -125,7 +146,7 @@ class AveragePrecision(Metric):
         target = dim_zero_cat(self.target)
         if not self.num_classes:
             raise ValueError(f"`num_classes` bas to be positive number, but got {self.num_classes}")
-        return _average_precision_compute(preds, target, self.num_classes, self.pos_label)
+        return _average_precision_compute(preds, target, self.num_classes, self.pos_label, self.average)
 
     @property
     def is_differentiable(self) -> bool: