+++ torchmetrics/functional/classification/average_precision.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from typing import List, Optional, Sequence, Tuple, Union
+import warnings
 
 import torch
 from torch import Tensor
@@ -27,8 +28,30 @@ def _average_precision_update(
     target: Tensor,
     num_classes: Optional[int] = None,
     pos_label: Optional[int] = None,
+    average: Optional[str] = 'macro',
 ) -> Tuple[Tensor, Tensor, int, Optional[int]]:
-    return _precision_recall_curve_update(preds, target, num_classes, pos_label)
+    """ Format the predictions and target based on the num_classes, pos_label and average parameter 
+    Args:
+        preds: predictions from model (logits or probabilities)
+        target: ground truth values
+        num_classes: integer with number of classes.
+        pos_label: integer determining the positive class. Default is ``None``
+            which for binary problem is translate to 1. For multiclass problems
+            this argument should not be set as we iteratively change it in the
+            range [0,num_classes-1]
+        average: reduction method for multi-class or multi-label problems
+    """
+    preds, target, num_classes, pos_label = _precision_recall_curve_update(preds, target, num_classes, pos_label)
+    if average == 'micro':
+        if preds.ndim == target.ndim:
+            # Considering each element of the label indicator matrix as a label
+            preds = preds.flatten()
+            target = target.flatten()
+            num_classes = 1
+        else:
+            raise ValueError('Cannot use `micro` average with multi-class input')
+
+    return preds, target, num_classes, pos_label
 
 
 def _average_precision_compute(
@@ -36,6 +59,7 @@ def _average_precision_compute(
     target: Tensor,
     num_classes: int,
     pos_label: Optional[int] = None,
+    average: Optional[str] = 'macro',
     sample_weights: Optional[Sequence] = None,
 ) -> Union[List[Tensor], Tensor]:
     """Computes the average precision score.
@@ -48,6 +72,7 @@ def _average_precision_compute(
             which for binary problem is translate to 1. For multiclass problems
             this argument should not be set as we iteratively change it in the
             range [0,num_classes-1]
+        average: reduction method for multi-class or multi-label problems
         sample_weights: sample weights for each data point
 
     Example:
@@ -67,19 +92,29 @@ def _average_precision_compute(
         >>> target = torch.tensor([0, 1, 3, 2])
         >>> num_classes = 5
         >>> preds, target, num_classes, pos_label = _average_precision_update(preds, target, num_classes)
-        >>> _average_precision_compute(preds, target, num_classes)
+        >>> _average_precision_compute(preds, target, num_classes, average=None)
         [tensor(1.), tensor(1.), tensor(0.2500), tensor(0.2500), tensor(nan)]
     """
 
     # todo: `sample_weights` is unused
     precision, recall, _ = _precision_recall_curve_compute(preds, target, num_classes, pos_label)
-    return _average_precision_compute_with_precision_recall(precision, recall, num_classes)
+    if average == 'weighted':
+        if preds.ndim == target.ndim and target.ndim > 1:
+            weights = target.sum(dim=0)
+        else:
+            weights = torch.bincount(target, minlength=num_classes)
+        weights = weights / sum(weights)
+    else:
+        weights = None
+    return _average_precision_compute_with_precision_recall(precision, recall, num_classes, average, weights)
 
 
 def _average_precision_compute_with_precision_recall(
     precision: Tensor,
     recall: Tensor,
     num_classes: int,
+    average: Optional[str] = 'macro',
+    weights: Optional[Tensor] = None,
 ) -> Union[List[Tensor], Tensor]:
     """Computes the average precision score from precision and recall.
 
@@ -88,6 +123,8 @@ def _average_precision_compute_with_precision_recall(
         recall: recall values
         num_classes: integer with number of classes. Not nessesary to provide
             for binary problems.
+        average: reduction method for multi-class or multi-label problems
+        weights: weights to use when average='weighted'
 
     Example:
         >>> # binary case
@@ -96,7 +133,7 @@ def _average_precision_compute_with_precision_recall(
         >>> pos_label = 1
         >>> preds, target, num_classes, pos_label = _average_precision_update(preds, target, pos_label=pos_label)
         >>> precision, recall, _ = _precision_recall_curve_compute(preds, target, num_classes, pos_label)
-        >>> _average_precision_compute_with_precision_recall(precision, recall, num_classes)
+        >>> _average_precision_compute_with_precision_recall(precision, recall, num_classes, average=None)
         tensor(1.)
 
         >>> # multiclass case
@@ -108,7 +145,7 @@ def _average_precision_compute_with_precision_recall(
         >>> num_classes = 5
         >>> preds, target, num_classes, pos_label = _average_precision_update(preds, target, num_classes)
         >>> precision, recall, _ = _precision_recall_curve_compute(preds, target, num_classes)
-        >>> _average_precision_compute_with_precision_recall(precision, recall, num_classes)
+        >>> _average_precision_compute_with_precision_recall(precision, recall, num_classes, average=None)
         [tensor(1.), tensor(1.), tensor(0.2500), tensor(0.2500), tensor(nan)]
     """
 
@@ -121,7 +158,24 @@ def _average_precision_compute_with_precision_recall(
     res = []
     for p, r in zip(precision, recall):
         res.append(-torch.sum((r[1:] - r[:-1]) * p[:-1]))
-    return res
+
+    # Reduce
+    if average in ('macro', 'weighted'):
+        res = torch.stack(res)
+        if torch.isnan(res).any():
+            warnings.warn('Average precision score for one or more classes was `nan`.'
+                          ' Ignoring these classes in average', UserWarning)
+        if average == 'macro':
+            return res[~torch.isnan(res)].mean()
+        else:
+            weights = torch.ones_like(res) if weights is None else weights
+            return (res * weights)[~torch.isnan(res)].sum()
+    elif average is None:
+        return res
+    else:
+        allowed_average = ('micro', 'macro', 'weighted', None)
+        raise ValueError(f'Expected argument `average` to be one of {allowed_average}'
+                         f' but got {average}')
 
 
 def average_precision(
@@ -129,6 +183,7 @@ def average_precision(
     target: Tensor,
     num_classes: Optional[int] = None,
     pos_label: Optional[int] = None,
+    average: Optional[str] = 'macro',
     sample_weights: Optional[Sequence] = None,
 ) -> Union[List[Tensor], Tensor]:
     """Computes the average precision score.
@@ -142,6 +197,21 @@ def average_precision(
             which for binary problem is translate to 1. For multiclass problems
             this argument should not be set as we iteratively change it in the
             range [0,num_classes-1]
+        average:
+            defines the reduction that is applied in the case of multiclass and multilabel input. 
+            Should be one of the following:
+
+            - ``'macro'`` [default]: Calculate the metric for each class separately, and average the
+              metrics across classes (with equal weights for each class).
+            - ``'micro'``: Calculate the metric globally, across all samples and classes. Cannot be
+              used with multiclass input.
+            - ``'weighted'``: Calculate the metric for each class separately, and average the
+              metrics across classes, weighting each class by its support.
+            - ``'none'`` or ``None``: Calculate the metric for each class separately, and return
+              the metric for every class.
+            - ``'samples'``: Calculate the metric for each sample, and average the metrics
+              across samples (with equal weights for each sample).
+
         sample_weights: sample weights for each data point
 
     Returns:
@@ -161,9 +231,9 @@ def average_precision(
         ...                      [0.05, 0.05, 0.75, 0.05, 0.05],
         ...                      [0.05, 0.05, 0.05, 0.75, 0.05]])
         >>> target = torch.tensor([0, 1, 3, 2])
-        >>> average_precision(pred, target, num_classes=5)
+        >>> average_precision(pred, target, num_classes=5, average=None)
         [tensor(1.), tensor(1.), tensor(0.2500), tensor(0.2500), tensor(nan)]
     """
     # fixme: `sample_weights` is unused
-    preds, target, num_classes, pos_label = _average_precision_update(preds, target, num_classes, pos_label)
-    return _average_precision_compute(preds, target, num_classes, pos_label, sample_weights)
+    preds, target, num_classes, pos_label = _average_precision_update(preds, target, num_classes, pos_label, average)
+    return _average_precision_compute(preds, target, num_classes, pos_label, average, sample_weights)